Make the inner classes of JettyHttpClient top-level
